Comment Auth

+++ Server/AppService/Application/Items/Commands/DeleteCommentCommand.cs
@@ -12,10 +12,12 @@ public record DeleteCommentCommand(string CommentId) : IRequest
     public class DeleteCommentCommandHandler : IRequestHandler<DeleteCommentCommand>
     {
         private readonly ICatalogContext context;
+        private readonly ICurrentUserService _currentUserService;
 
-        public DeleteCommentCommandHandler(ICatalogContext context)
+        public DeleteCommentCommandHandler(ICatalogContext context, ICurrentUserService currentUserService)
         {
             this.context = context;
+            _currentUserService = currentUserService;
         }
 
         public async Task<Unit> Handle(DeleteCommentCommand request, CancellationToken cancellationToken)
@@ -26,13 +28,20 @@ public record DeleteCommentCommand(string CommentId) : IRequest
 
             if (comment is null) throw new Exception();
 
+            if (!IsAuthorizedToDelete(comment)) 
+            {
+                throw new UnauthorizedAccessException("Unauthorized");
+            }
+
             context.Comments.Remove(comment);
 
             comment.Item.CommentCount--;
-           
+
             await context.SaveChangesAsync(cancellationToken);
 
             return Unit.Value;
         }
+
+        private bool IsAuthorizedToDelete(Domain.Entities.Comment comment) => _currentUserService.IsCurrentUser(comment.CreatedById!) || _currentUserService.IsUserInRole(Roles.Administrator);
     }
 }
